requirements.txt
selenium
webdriver-manager
beautifulsoup4
psycopg2-binary
requests
